# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Master Key Consulting API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Review {
  id: number;
  name: string;
  location: string;
  rating: number;
  comment: string;
  service: string;
  approved: boolean;
  createdAt: string;
}

export interface ReviewInput {
  name: string;
  location: string;
  rating: number;
  comment: string;
  service: string;
}

export interface ReviewUpdate {
  approved?: boolean;
}

export interface Project {
  id: number;
  title: string;
  location: string;
  serviceType: string;
  description: string;
  date: string;
  /** @nullable */
  imageUrl?: string | null;
  featured: boolean;
  createdAt: string;
}

export interface ProjectInput {
  title: string;
  location: string;
  serviceType: string;
  description: string;
  date: string;
  imageUrl?: string;
  featured?: boolean;
}

export interface ProjectUpdate {
  title?: string;
  location?: string;
  serviceType?: string;
  description?: string;
  date?: string;
  imageUrl?: string;
  featured?: boolean;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  content: string;
  /** @nullable */
  coverImageUrl?: string | null;
  tags: string[];
  published: boolean;
  createdAt: string;
}

export interface BlogPostInput {
  title: string;
  content: string;
  coverImageUrl?: string;
  tags?: string[];
  published?: boolean;
}

export interface BlogPostUpdate {
  title?: string;
  content?: string;
  coverImageUrl?: string;
  tags?: string[];
  published?: boolean;
}

export interface Booking {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  location: string;
  serviceType: string;
  preferredDate: string;
  message: string;
  /** @nullable */
  howHeard?: string | null;
  status: string;
  createdAt: string;
}

export interface BookingInput {
  fullName: string;
  email: string;
  phone: string;
  location: string;
  serviceType: string;
  preferredDate: string;
  message: string;
  howHeard?: string;
}

export interface BookingUpdate {
  status?: string;
}

export interface AdminStats {
  totalBookings: number;
  newBookings: number;
  publishedPosts: number;
  pendingReviews: number;
  totalProjects: number;
  totalReviews: number;
}

export interface AdminLoginInput {
  password: string;
}

export interface AdminLoginResult {
  authenticated: boolean;
}

export type ListProjectsParams = {
  serviceType?: string;
};