# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Master Key Consulting API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: reviews
    description: Client reviews
  - name: projects
    description: Portfolio projects
  - name: blog
    description: Blog posts
  - name: bookings
    description: Service bookings
  - name: admin
    description: Admin operations

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List approved reviews
      responses:
        "200":
          description: List of approved reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review (pending approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reviews/all:
    get:
      operationId: listAllReviews
      tags: [reviews]
      summary: List all reviews (admin)
      responses:
        "200":
          description: All reviews including unapproved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  /reviews/{id}:
    patch:
      operationId: updateReview
      tags: [reviews]
      summary: Update review (admin — approve/reject)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "404":
          description: Not found
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: serviceType
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/featured:
    get:
      operationId: listFeaturedProjects
      tags: [projects]
      summary: List featured projects for homepage
      responses:
        "200":
          description: Featured projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a single project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /blog:
    get:
      operationId: listBlogPosts
      tags: [blog]
      summary: List published blog posts
      responses:
        "200":
          description: Published blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
    post:
      operationId: createBlogPost
      tags: [blog]
      summary: Create a blog post (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostInput"
      responses:
        "201":
          description: Created blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  /blog/all:
    get:
      operationId: listAllBlogPosts
      tags: [blog]
      summary: List all blog posts including drafts (admin)
      responses:
        "200":
          description: All blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"

  /blog/{slug}:
    get:
      operationId: getBlogPost
      tags: [blog]
      summary: Get a blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found

  /blog/{id}/update:
    patch:
      operationId: updateBlogPost
      tags: [blog]
      summary: Update a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlogPostUpdate"
      responses:
        "200":
          description: Updated blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found

  /blog/{id}/delete:
    delete:
      operationId: deleteBlogPost
      tags: [blog]
      summary: Delete a blog post (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings (admin)
      responses:
        "200":
          description: All bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Submit a booking request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  /bookings/{id}:
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update booking status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingUpdate"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found

  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Get admin dashboard statistics
      responses:
        "200":
          description: Dashboard stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Invalid credentials

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Current admin info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResult"
        "401":
          description: Not authenticated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Review:
      type: object
      required: [id, name, location, rating, comment, service, approved, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        location:
          type: string
        rating:
          type: integer
        comment:
          type: string
        service:
          type: string
        approved:
          type: boolean
        createdAt:
          type: string

    ReviewInput:
      type: object
      required: [name, location, rating, comment, service]
      properties:
        name:
          type: string
        location:
          type: string
        rating:
          type: integer
        comment:
          type: string
        service:
          type: string

    ReviewUpdate:
      type: object
      properties:
        approved:
          type: boolean

    Project:
      type: object
      required: [id, title, location, serviceType, description, date, featured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        location:
          type: string
        serviceType:
          type: string
        description:
          type: string
        date:
          type: string
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        createdAt:
          type: string

    ProjectInput:
      type: object
      required: [title, location, serviceType, description, date]
      properties:
        title:
          type: string
        location:
          type: string
        serviceType:
          type: string
        description:
          type: string
        date:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean

    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
        location:
          type: string
        serviceType:
          type: string
        description:
          type: string
        date:
          type: string
        imageUrl:
          type: string
        featured:
          type: boolean

    BlogPost:
      type: object
      required: [id, title, slug, content, tags, published, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        content:
          type: string
        coverImageUrl:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean
        createdAt:
          type: string

    BlogPostInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean

    BlogPostUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        coverImageUrl:
          type: string
        tags:
          type: array
          items:
            type: string
        published:
          type: boolean

    Booking:
      type: object
      required: [id, fullName, email, phone, location, serviceType, preferredDate, message, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        serviceType:
          type: string
        preferredDate:
          type: string
        message:
          type: string
        howHeard:
          type: ["string", "null"]
        status:
          type: string
        createdAt:
          type: string

    BookingInput:
      type: object
      required: [fullName, email, phone, location, serviceType, preferredDate, message]
      properties:
        fullName:
          type: string
        email:
          type: string
        phone:
          type: string
        location:
          type: string
        serviceType:
          type: string
        preferredDate:
          type: string
        message:
          type: string
        howHeard:
          type: string

    BookingUpdate:
      type: object
      properties:
        status:
          type: string

    AdminStats:
      type: object
      required: [totalBookings, newBookings, publishedPosts, pendingReviews, totalProjects, totalReviews]
      properties:
        totalBookings:
          type: integer
        newBookings:
          type: integer
        publishedPosts:
          type: integer
        pendingReviews:
          type: integer
        totalProjects:
          type: integer
        totalReviews:
          type: integer

    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string

    AdminLoginResult:
      type: object
      required: [authenticated]
      properties:
        authenticated:
          type: boolean