# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Master Key Consulting API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  totalBookings: number;
  newBookings: number;
  publishedPosts: number;
  pendingReviews: number;
  totalProjects: number;
  totalReviews: number;
}