# lib/api-zod/src/generated/types/blogPost.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Master Key Consulting API
 * OpenAPI spec version: 0.1.0
 */

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  content: string;
  /** @nullable */
  coverImageUrl?: string | null;
  tags: string[];
  published: boolean;
  createdAt: string;
}