# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Master Key Consulting API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  fullName: string;
  email: string;
  phone: string;
  location: string;
  serviceType: string;
  preferredDate: string;
  message: string;
  howHeard?: string;
}