# lib/api-zod/src/generated/types/projectInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Master Key Consulting API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectInput {
  title: string;
  location: string;
  serviceType: string;
  description: string;
  date: string;
  imageUrl?: string;
  featured?: boolean;
}